--- src/lib/finders/index.ts
@@ -1,12 +1,0 @@
-/**
- * Central exports for all finder functions
- * Enables clean single-line imports: import { findModelInTargetInstance, findPageInTargetInstance, ... } from './lib/finders'
- */
-
-export { findAssetInTargetInstance } from './asset-finder';
-export { findContainerInTargetInstance, findContainerInTargetInstanceEnhanced } from './container-finder';
-export { findContentInTargetInstance, findContentInTargetInstanceLegacy } from './content-item-finder';
-export { findGalleryInTargetInstance } from './gallery-finder';
-export { findModelInTargetInstance, findModelInTargetInstanceEnhanced } from './model-finder';
-export { findPageInTargetInstance, findPageInTargetInstanceEnhanced } from './page-finder';
-export { findTemplateInTargetInstanceEnhanced } from './template-finder';
